+++ src/fidesops/ops/service/connectors/__init__.py
@@ -5,6 +5,9 @@ from fidesops.ops.service.connectors.base_connector import BaseConnector
 from fidesops.ops.service.connectors.email_connector import EmailConnector
 from fidesops.ops.service.connectors.http_connector import HTTPSConnector
 from fidesops.ops.service.connectors.manual_connector import ManualConnector
+from fidesops.ops.service.connectors.manual_webhook_connector import (
+    ManualWebhookConnector,
+)
 from fidesops.ops.service.connectors.mongodb_connector import MongoDBConnector
 from fidesops.ops.service.connectors.saas_connector import SaaSConnector
 from fidesops.ops.service.connectors.sql_connector import (
@@ -30,6 +33,7 @@ supported_connectors: Dict[str, Any] = {
     ConnectionType.bigquery.value: BigQueryConnector,
     ConnectionType.manual.value: ManualConnector,
     ConnectionType.email.value: EmailConnector,
+    ConnectionType.manual_webhook.value: ManualWebhookConnector,
 }